+++ src/jk_version/__init__.py
@@ -2,7 +2,7 @@
 
 
 __author__ = "Jürgen Knauth"
-__version__ = "0.2022.8.18"
+__version__ = "0.2023.7.11"